requirements.txt
streamlit==1.43.2
streamlit-chat==0.0.2.2
openai==1.3.0
python-dotenv==1.0.0
Pillow>=9.1.0,<11.0.0